EVM-161 Updated Web3J and implemented EthereumTransaction with it. Also added semantic validator and updated tests

+++ sdk/src/main/scala/com/horizen/params/NetworkParams.scala
@@ -63,5 +63,5 @@ trait NetworkParams {
   val allowedForgersList: Seq[(PublicKey25519Proposition, VrfPublicKey)] = Seq()
 
   // Account chain params
-  val chainId : Int
+  val chainId : Long
 }
